Fixed issue which caused .demand function not to work correctly.
The code for copying attributes of an optimist instance to the module had a
flaw which resulted in the demand function not working. Specifically, the
"argv" special property was having its value copied over, instead of having its
accessor method proxied. Therefore, .argv would always contain the initial
output of parseArgs(processArgs), that is, the output before any .demand calls
have been made.

The fix consists of proxying the argv accessor in the module object.